requirements.txt
Flask
Flask-Cors
boto3
opencv-python
opencv-contrib-python
torch
torchvision
ultralytics
pandas
requests
numpy
Pillow
tqdm
scikit-learn
fastapi
pydantic
uvicorn
tensorflow
gunicorn
gitpython
